refactor toggler to improve JS
