+++ src/assets/js/toggler.js
@@ -1,0 +1,48 @@
+/**
+ * Add toggles for sections in the sidebar on /rest/reference.
+ *
+ * @author Tim Scanlin (tim.scanlin@optimizely.com)
+ */
+
+module.exports = function() {
+  var each = [].forEach;
+  var body = document.body;
+
+  // Constants.
+  var TOGGLE_CONTAINER_CLASS = 'js-toggle-container';
+  var TOGGLE_SECTION_ATTRIBUTE = 'data-toggle-section';
+  var TOGGLE_TRIGGER_ATTRIBUTE = 'data-toggle-trigger';
+  var TOGGLE_TRIGGER_CLASS_ATTRIBUTE = 'data-toggle-trigger-class';
+  var TOGGLE_CLASS = 'visible';
+
+  var toggleContainer = body.querySelector('.' + TOGGLE_CONTAINER_CLASS);
+
+  if (toggleContainer !== null) {
+    var toggleSections = body.querySelectorAll('[' + TOGGLE_SECTION_ATTRIBUTE + ']');
+    var toggleTriggers = toggleContainer.querySelectorAll('[' + TOGGLE_TRIGGER_ATTRIBUTE + ']');
+    each.call(toggleTriggers, function(elm) {
+      elm.addEventListener('click', function(e) {
+        e.preventDefault();
+        var target = e.target;
+        var toggleTriggerClass = target.getAttribute(TOGGLE_TRIGGER_CLASS_ATTRIBUTE);
+        var toggleSectionId = target.getAttribute(TOGGLE_TRIGGER_ATTRIBUTE);
+
+        each.call(toggleContainer.querySelectorAll('.' + toggleTriggerClass), function(el) {
+          el.classList.remove(toggleTriggerClass);
+        });
+        target.classList.add(toggleTriggerClass);
+
+        // Remove visible class from all toggle sections.
+        each.call(toggleSections, function(el) {
+          el.classList.remove(TOGGLE_CLASS);
+        });
+
+        // Add visible class to toggle sections that have a matching id.
+        var sectionsToToggle = body.querySelectorAll('[' + TOGGLE_SECTION_ATTRIBUTE + '="' + toggleSectionId + '"]');
+        each.call(sectionsToToggle, function(el) {
+          el.classList.add(TOGGLE_CLASS);
+        });
+      });
+    });
+  }
+};
