+++ src/assets/scss/_base.scss
@@ -72,10 +72,6 @@ em, b, strong {
   }
 }
 
-.hidden {
-  display: none;
-}
-
 .white {
   color: white;
 }
