+++ src/assets/scss/_state.scss
@@ -16,6 +16,10 @@
   max-height: 0;
 }
 
+.hidden {
+  display: none;
+}
+
 .visible {
   display: block;
   visibility: visible;
